[ft-template] update doc links (#294)
# What does this PR do?

Updates links to documentation in `llmforge` . Also, I noticed that
there are merge conflicts between `release_branch` and `main` in #292
for reasons I don't completely understand (we've always kept this branch
to track `main` directly). The only way to fix this is to merge changes
from release branch into `main` and then merge `main` into the release
branch. So this PR also does that.