Add Deepseek R1 example to ray_serve_llm template (#451)
Adds a tutorial for launching DeepSeek R1 using an autoscaling cluster.
For the purposes of this PR, I am not including a service.yaml for
`serve run` because we have encountered a few errors and we need this
template out before the fix will be ready.

cc @kevin85421 @nrghosh @eicherseiji

---------

Co-authored-by: Seiji Eicher <58963096+eicherseiji@users.noreply.github.com>