Add deprecation notice (#501)
### Summary

The post-summit process will be - all templates get committed in the ray
repo with proper testing. These will be automatically released through
the new publishing workflow that Lonnie is working on. There are some
templates in anyscale/templates repo that will need to be migrated.

Ray Summit 2025 templates are an exception (they will be in
anyscale/templates repo, and not tested) - they will be removed right
after the summit.

Here we add a deprecation notice to make it clear that all updates or
new templates should be done in Ray repo.

Co-authored-by: Gang Zhao <gang@gang-JQ62HD2C37.local>